move proving inputs from app to register inputs in common
